multipack support for official mixtral implementation
