+++ requirements.txt
@@ -2,7 +2,7 @@
 auto-gptq==0.5.1
 packaging
 peft==0.6.0
-transformers @ git+https://github.com/huggingface/transformers.git@df5c5c62ae253055336f5bb0828ca8e3e15ab6bd
+transformers==4.36.0
 tokenizers==0.15.0
 bitsandbytes>=0.41.1
 accelerate==0.24.1
